Feature02/profile & mypage

+++ src/main/java/com/beteam/willu/user/controller/UserController.java
@@ -1,5 +1,8 @@
 package com.beteam.willu.user.controller;
 
+import java.util.Map;
+
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
@@ -64,7 +67,14 @@ public class UserController {
 
 	@DeleteMapping("/users/{id}")
 	public ResponseEntity<ApiResponseDto> deleteUser(@PathVariable Long id,
-		@AuthenticationPrincipal UserDetailsImpl userDetails) {
+		@RequestBody Map<String, String> request, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+		String password = request.get("password");
+
+		// 사용자의 비밀번호 확인
+		if (!userService.isPasswordCorrect(userDetails.getUser().getId(), password)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ApiResponseDto("비밀번호가 일치하지 않습니다.", 401));
+		}
+
 		userService.deleteUser(id, userDetails.getUser());
 		return ResponseEntity.ok(new ApiResponseDto("회원 탈퇴 성공", 200));
 	}
